# deploy/demo/openapi/specs/openapi-localhost:9082.yml
consumes:
- application/json
produces:
- application/json
schemes:
- https
swagger: "2.0"
info:
  title: (Aries Agent REST Server) of aries-framework-go.
  license:
    name: 'SPDX-License-Identifier: Apache-2.0'
  version: 0.1.0
host: localhost:9082
paths:
...
  /vcwallet/add:
    post:
      description: |-
        Supported data models:
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Collection
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Credential
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#DIDResolutionResponse
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#meta-data
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#connection
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Key
      tags:
      - vcwallet
      summary: adds given data model to wallet content store.
      operationId: addContentReq
      parameters:
      - description: Params for adding content to wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/AddContentRequest'
      responses:
        "200":
          $ref: '#/responses/emptyRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/close:
    post:
      description: returns response containing bool flag false if token is not found
        or already expired for this wallet user.
      tags:
      - vcwallet
      summary: Expires token issued to this VC wallet, removes wallet's key manager
        instance and closes wallet content store.
      operationId: lockWalletReq
      parameters:
      - description: Params for locking wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/LockWalletRequest'
      responses:
        "200":
          $ref: '#/responses/lockWalletRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/connect:
    post:
      tags:
      - vcwallet
      summary: accepts out-of-band invitations and performs DID exchange.
      operationId: connectReq
      parameters:
      - description: Params for connecting to wallet for DIDComm.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/ConnectRequest'
      responses:
        "200":
          $ref: '#/responses/connectRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/create-key-pair:
    post:
      tags:
      - vcwallet
      summary: creates a new key pair from wallet.
      operationId: createKeyPairReq
      parameters:
      - description: Params for creating key pair from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/CreateKeyPairRequest'
      responses:
        "200":
          $ref: '#/responses/createKeyPairRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/create-profile:
    post:
      tags:
      - vcwallet
      summary: Creates new wallet profile and returns error if wallet profile is already
        created.
      operationId: createProfileReq
      responses:
        "200":
          $ref: '#/responses/emptyRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/derive:
    post:
      description: https://w3c-ccg.github.io/universal-wallet-interop-spec/#derive
      tags:
      - vcwallet
      summary: derives a Verifiable Credential.
      operationId: deriveReq
      parameters:
      - description: Params for deriving a credential from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/DeriveRequest'
      responses:
        "200":
          $ref: '#/responses/deriveRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/get:
    post:
      description: |-
        Supported data models:
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Collection
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Credential
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#DIDResolutionResponse
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#meta-data
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#connection
      tags:
      - vcwallet
      summary: gets content from wallet content store.
      operationId: getContentReq
      parameters:
      - description: Params for getting content from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/GetContentRequest'
      responses:
        "200":
          $ref: '#/responses/getContentRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/getall:
    post:
      description: |-
        Supported data models:
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Collection
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Credential
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#DIDResolutionResponse
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#meta-data
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#connection
      tags:
      - vcwallet
      summary: gets all contents from wallet content store for given content type.
      operationId: getAllContentReq
      parameters:
      - description: Params for getting all contents from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/GetAllContentRequest'
      responses:
        "200":
          $ref: '#/responses/getAllContentRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/issue:
    post:
      description: https://w3c-ccg.github.io/universal-wallet-interop-spec/#issue
      tags:
      - vcwallet
      summary: adds proof to a Verifiable Credential.
      operationId: issueReq
      parameters:
      - description: Params for issuing credential from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/IssueRequest'
      responses:
        "200":
          $ref: '#/responses/issueRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/open:
    post:
      description: |-
        Unlocks given wallet's key manager instance & content store and
        returns a authorization token to be used for performing wallet operations.
      tags:
      - vcwallet
      operationId: unlockWalletReq
      parameters:
      - description: Params for unlocking wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/UnlockWalletRequest'
      responses:
        "200":
          $ref: '#/responses/unlockWalletRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/present-proof:
    post:
      description: |-
        Currently Supporting
        [0454-present-proof-v2](https://github.com/hyperledger/aries-rfcs/tree/master/features/0454-present-proof-v2)
      tags:
      - vcwallet
      summary: |-
        sends message present proof message from wallet to relying party.
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#presentproof
      operationId: presentProofReq
      parameters:
      - description: Params for accepting presentation request and sending present
          proof message to relying party.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/PresentProofRequest'
      responses:
        "200":
          $ref: '#/responses/presentProofRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/profile/{id}:
    get:
      tags:
      - vcwallet
      summary: Checks if profile exists for given wallet user ID and returns error
        if profile doesn't exists.
      operationId: checkProfile
      parameters:
      - type: string
        x-go-name: ID
        description: Wallet User's ID used to create profile.
        name: id
        in: path
        required: true
      responses:
        "200":
          $ref: '#/responses/emptyRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/propose-credential:
    post:
      description: |-
        Currently Supporting : 0453-issueCredentialV2
        https://github.com/hyperledger/aries-rfcs/blob/main/features/0453-issue-credential-v2/README.md
      tags:
      - vcwallet
      summary: |-
        Sends propose credential message from wallet to issuer and optionally waits for offer credential response.
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#proposecredential
      operationId: proposeCredReq
      parameters:
      - description: Params for proposing credential from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/ProposeCredentialRequest'
      responses:
        "200":
          $ref: '#/responses/proposeCredRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/propose-presentation:
    post:
      description: |-
        Currently Supporting
        [0454-present-proof-v2](https://github.com/hyperledger/aries-rfcs/tree/master/features/0454-present-proof-v2)
      tags:
      - vcwallet
      summary: |-
        accepts out-of-band invitation and sends message proposing presentation
        from wallet to relying party.
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#proposepresentation
      operationId: proposePresReq
      parameters:
      - description: Params for proposing presentation from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/ProposePresentationRequest'
      responses:
        "200":
          $ref: '#/responses/proposePresRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/prove:
    post:
      description: https://w3c-ccg.github.io/universal-wallet-interop-spec/#prove
      tags:
      - vcwallet
      summary: produces a Verifiable Presentation.
      operationId: proveReq
      parameters:
      - description: Params for producing verifiable presentation from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/ProveRequest'
      responses:
        "200":
          $ref: '#/responses/proveRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/query:
    post:
      description: |-
        This function may return multiple presentations as a result based on combination of query types used.

        https://w3c-ccg.github.io/universal-wallet-interop-spec/#query

        Supported Query Types:
        https://www.w3.org/TR/json-ld11-framing
        https://identity.foundation/presentation-exchange
        https://w3c-ccg.github.io/vp-request-spec/#query-by-example
        https://w3c-ccg.github.io/vp-request-spec/#did-authentication-request
      tags:
      - vcwallet
      summary: runs query against wallet credential contents and returns presentation
        containing credential results.
      operationId: contentQueryReq
      parameters:
      - description: Params for querying credentials from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/ContentQueryRequest'
      responses:
        "200":
          $ref: '#/responses/contentQueryRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/remove:
    post:
      description: |-
        Supported data models:
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Collection
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#Credential
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#DIDResolutionResponse
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#meta-data
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#connection
      tags:
      - vcwallet
      summary: removes given content from wallet content store.
      operationId: removeContentReq
      parameters:
      - description: Params for removing content from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/RemoveContentRequest'
      responses:
        "200":
          $ref: '#/responses/emptyRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/request-credential:
    post:
      description: |-
        Currently Supporting : 0453-issueCredentialV2
        https://github.com/hyperledger/aries-rfcs/blob/main/features/0453-issue-credential-v2/README.md
      tags:
      - vcwallet
      summary: |-
        Sends request credential message from wallet to issuer and optionally waits for credential response.
        https://w3c-ccg.github.io/universal-wallet-interop-spec/#requestcredential
      operationId: requestCredReq
      parameters:
      - description: Params for sending request credential message from wallet and
          optionally wait for credential response.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/RequestCredentialRequest'
      responses:
        "200":
          $ref: '#/responses/requestCredRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/resolve-credential-manifest:
    post:
      description: 'Supports: https://identity.foundation/credential-manifest/'
      tags:
      - vcwallet
      summary: Resolves given credential manifest by credential response or credential.
      operationId: resolveCredManifest
      parameters:
      - description: Params for resolving credential manifests from wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/ResolveCredentialManifestRequest'
      responses:
        "200":
          $ref: '#/responses/resolveCredManifest'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/update-profile:
    post:
      description: |-
        Caution:
        you might lose your existing keys if you change kms options.
        you might lose your existing wallet contents if you change storage/EDV options
        (example: switching context storage provider or changing EDV settings).
      tags:
      - vcwallet
      summary: Updates an existing wallet profile and returns error if profile doesn't
        exists.
      operationId: UpdateProfileReq
      responses:
        "200":
          $ref: '#/responses/emptyRes'
        default:
          $ref: '#/responses/genericError'
  /vcwallet/verify:
    post:
      description: https://w3c-ccg.github.io/universal-wallet-interop-spec/#prove
      tags:
      - vcwallet
      summary: verifies a Verifiable Credential or a Verifiable Presentation.
      operationId: verifyReq
      parameters:
      - description: Params for producing verifying a credential or presentation from
          wallet.
        name: Params
        in: body
        schema:
          $ref: '#/definitions/VerifyRequest'
      responses:
        "200":
          $ref: '#/responses/verifyRes'
        default:
          $ref: '#/responses/genericError'
...
  /verifiable/presentations:
    get:
      summary: Retrieves the verifiable credentials.
      operationId: verifiable
      responses:
        "200":
          $ref: '#/responses/presentationRecordResult'
        default:
          $ref: '#/responses/genericError'
...
definitions:
  AddContentRequest:
    type: object
    title: AddContentRequest is request for adding a content to wallet.
    properties:
      auth:
        description: Authorization token for performing wallet operations.
        type: string
        x-go-name: Auth
      collectionID:
        description: ID of the wallet collection to which this content should belong.
        type: string
        x-go-name: CollectionID
      content:
        description: content to be added to wallet content store.
        type: object
        x-go-name: Content
      contentType:
        $ref: '#/definitions/ContentType'
      userID:
        description: ID of wallet user.
        type: string
        x-go-name: UserID
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/command/vcwallet
...
  ConnectRequest:
    type: object
    title: ConnectRequest is request model for wallet DID connect operation.
    properties:
      auth:
        description: Authorization token for performing wallet operations.
        type: string
        x-go-name: Auth
      invitation:
        $ref: '#/definitions/Invitation'
      myLabel:
        description: Label to be shared with the other agent during the subsequent
          DID exchange.
        type: string
        x-go-name: MyLabel
      reuseAnyConnection:
        description: To use any recognized DID in the services array for a reusable
          connection.
        type: boolean
        x-go-name: ReuseAnyConnection
      reuseConnection:
        description: DID to be used when reusing a connection.
        type: string
        x-go-name: ReuseConnection
      routerConnections:
        description: router connections to be used to establish connection.
        type: array
        items:
          type: string
        x-go-name: RouterConnections
      timeout:
        $ref: '#/definitions/Duration'
      userID:
        description: ID of wallet user.
        type: string
        x-go-name: UserID
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/command/didcommwallet
...
  CreateKeyPairRequest:
    type: object
    title: CreateKeyPairRequest is request model for creating key pair from wallet.
    properties:
      attrs:
        description: extra options needed for creating some keys
        type: array
        items:
          type: string
        x-go-name: Attrs
      auth:
        description: Authorization token for performing wallet operations.
        type: string
        x-go-name: Auth
      keyType:
        $ref: '#/definitions/KeyType'
      userID:
        description: ID of wallet user.
        type: string
        x-go-name: UserID
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/command/vcwallet
...
  LockWalletRequest:
    type: object
    title: LockWalletRequest contains options for locking wallet.
    properties:
      userID:
        description: user ID of the wallet to be locked.
        type: string
        x-go-name: UserID
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/command/vcwallet
...
responses:
...
  connectRes:
    description: connectResponse is response model from wallet DID connect operation.
    schema:
      $ref: '#/definitions/ConnectResponse'
...
  emptyRes:
    description: emptyRes model
...
  genericError:
    description: genericError is aries rest api error response
    schema:
      $ref: '#/definitions/genericErrorBody'
...
  lockWalletRes:
    description: lockWalletResponse contains response for wallet lock operation.
    headers:
      closed:
        type: boolean
        description: |-
          Closed status of the wallet lock operation.
          if true, wallet is closed successfully
          if false, wallet is already closed or never unlocked.
...